# docs/swagger.yaml
basePath: /
definitions:
  handler.loginRequest:
    properties:
      password:
        example: Admin@123456
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  handler.refreshRequest:
    properties:
      refresh_token:
        example: refresh-token
        type: string
    required:
    - refresh_token
    type: object
  handler.registerRequest:
    properties:
      email:
        example: admin@example.com
        type: string
      nickname:
        example: admin
        type: string
      password:
        example: Admin@123456
        type: string
      username:
        example: admin
        type: string
    required:
    - nickname
    - password
    - username
    type: object
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Gin Rocket backend APIs with RBAC and JWT authentication.
  title: Gin Rocket API
  version: 1.0.0
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login with username and password, then return tokens, user info
        and permission codes.
      parameters:
      - description: login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      summary: Login
      tags:
      - auth
  /api/v1/auth/me:
    get:
      consumes:
      - application/json
      description: Return current user profile and permission code list for menu and
        button rendering.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange refresh token for a new access token and refresh token.
      parameters:
      - description: refresh request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      summary: Refresh token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user. The first registered user becomes super-admin
        automatically.
      parameters:
      - description: register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      summary: Register user
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: 'Use format: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"